+++ nbgrader/auth/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseAuth
+from .hubauth import HubAuth
+from .noauth import NoAuth
